+++ react-components/src/components/ProductFinder/CountryFinderModal.jsx
@@ -189,15 +189,30 @@ export default function CountryFinderModal(props) {
       <div className="suggested-markets">
         <h3 className="h-s">
           Suggested markets
-          {suggestedCountries.hs2Desc &&
-            ` for ${suggestedCountries.hs2Desc}`}
+          {suggestedCountries.hs2Desc && ` for ${suggestedCountries.hs2Desc}`}
         </h3>
-        <p className="m-v-xs">These are based on the size of the market for your product, export distance, tariffs and costs.</p>
+        <p className="m-v-xs">
+          These markets are based on consumer demand, export distance, tariffs
+          and costs for exporting{' '}
+          <span className="body-l-b">
+            {suggestedCountries.hs2Desc.toLowerCase()}
+          </span>{' '}
+          from the UK. This is an HS2 category that includes{' '}
+          <span className="body-l-b">
+            {suggestedCountries.details.product} (HS code{' '}
+            {suggestedCountries.details.hs6})
+          </span>
+          , along with other products categorised at an HS4 and HS6 level. For
+          more information on HS codes see our{' '}
+          <a href="/learn/categories/selling-across-borders-product-and-services-regulations-licensing-and-logistics/get-your-goods-into-the-destination-country/using-commodity-codes/">
+            lesson on using HS codes
+          </a>
+          .
+        </p>
         <div className="m-v-xs">{suggestedList}</div>
       </div>
     )
   }
-
   /*   Compare markets section  */
   const compareMarketsSection = (
     <div>